Improvements to intefaces and results
Print the string version of the Result to the log after executing, also
change the Internal Error result to render a Internal Server Error h1
tag. If you want a special Error, make your own. Also start keeping the
PreFilter/PreItem/SetContext ability of the controller as part of the
Leaf data, so we won't make interface assertions unless we have a good
idea that we should.